+++ src/main/java/com/br/avaliationsystemecommerce/dto/AvaliationRequest.java
@@ -5,10 +5,12 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record AvaliationRequest (
         @NotBlank @Size(max=155) String clientName,
         @Size(max=1500) String comment,
-        @Valid() Float avaliation,
+        @Valid() BigDecimal avaliation,
         @NotNull Long productId
 ) {
 }
